COMMERCE-2475 dispatch - move to new package (goal is to avoid using word scheduler)
