+++ modules/apps/dispatch/dispatch-api/bnd.bnd
@@ -2,9 +2,9 @@ Bundle-Name: Liferay Dispatch API
 Bundle-SymbolicName: com.liferay.dispatch.api
 Bundle-Version: 1.0.0
 Export-Package:\
+	com.liferay.dispatch.advisor,\
 	com.liferay.dispatch.constants,\
 	com.liferay.dispatch.exception,\
 	com.liferay.dispatch.model,\
-	com.liferay.dispatch.scheduler,\
 	com.liferay.dispatch.service,\
 	com.liferay.dispatch.service.persistence
