+++ modules/apps/dispatch/dispatch-api/src/main/resources/com/liferay/dispatch/advisor/packageinfo
